pyproject.toml
[project]
name = "churn-ml-pipeline"
version = "0.1.0"
description = "Production-ready churn prediction ML pipeline with training, validation, packaging, and serving."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115.0",
  "httpx>=0.27.0",
  "joblib>=1.4.2",
  "mlflow>=3.10.1",
  "numpy>=1.26.4",
  "pandas>=2.2.3",
  "pydantic>=2.9.2",
  "pyyaml>=6.0.2",
  "scikit-learn>=1.5.2",
  "uvicorn>=0.32.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.3",
]

[tool.pytest.ini_options]
pythonpath = [
  "src",
]
testpaths = [
  "tests",
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]
